Use 'float' types when testing powf
Using 'double' on i386 would end up using whatever double bits were in
the return value, instead of having that converted to float. Thanks i386.

+++ newlib/libm/test/math2.c
@@ -200,10 +200,10 @@ test_pow (void)
   
   for (i = 0; i < 1000; i++) 
   {
-    double n1;
-    double n2;
-    double res;
-    double shouldbe;
+    float n1;
+    float n2;
+    float res;
+    float shouldbe;
 
     errno = 0;
